Move type/use util function to patient area
