+++ apps/modernization-ui/src/address/display/displayAddress.spec.tsx
@@ -1,5 +1,5 @@
 import { ItemGroup } from 'design-system/item';
-import { displayAddress, displayAddressText, displayAddressTypeUse } from './displayAddress';
+import { displayAddress, displayAddressText } from './displayAddress';
 
 describe('when given an address', () => {
     it('should display the formatted address as a string', () => {
@@ -26,32 +26,5 @@ describe('when given an address', () => {
         const addressElement = displayAddress(address);
         expect(addressElement.type).toBe(ItemGroup);
         expect(addressElement.props.children).toBe('14 North Moore St, Suite A\nNew York, NY 10013');
-    });
-
-    it('should display the type and use as a string', () => {
-        const address = {
-            type: 'Dormitory',
-            use: 'Home'
-        };
-
-        expect(displayAddressTypeUse(address)).toBe('Dormitory/Home');
-    });
-
-    it('should display only the type when use is null', () => {
-        const address = {
-            type: 'Dormitory',
-            use: null
-        };
-
-        expect(displayAddressTypeUse(address)).toBe('Dormitory');
-    });
-
-    it('should display empty string when type and use are null', () => {
-        const address = {
-            type: null,
-            use: null
-        };
-
-        expect(displayAddressTypeUse(address)).toBe('');
     });
 });
